Many Changes:
Ok, I got through the code and found some problems, which I fixed.
1. I renamed the folder from 'pas' to 'patchprotect'
2. Fixed some bugs in PropProtection. There was a problem by setting the
owner
3. Fixed some other cantool/canpickup permissions. The if condition was
too complicated and didn't work well. Now it is much better and smarter
4. Improoved pas.lua very much. Now it also cooperates with npcs and
property permissions.
5. Added CanDrive to propprotection

Anyway, there have to be more things added to prop protection, like
buddies and some options to set. But for now it works good.

I tested all that on my server quite often and fixed all problems which
I created with the changes. If there is still an error, give me a shout!