Fixed an important bug
There was an exploit in the if cleanup condition.
the return if's didn't send the constraint-entities to the CPPISetOwner
function. For that, the function for checking constraints didn't work.
Also I added an exception that the prop, where f.e. a thruster is
spawned, doesn't override the owner of the current entity

Special thanks to Broteuss