Fixed a tricky glitch
First of all I have to say that this bug is not only in PatchProtect. It
is in nearly all PP's which are existing for GMod.

The problem is, that people use the "/../" to prevent the prop-patch
check to be able to spawn blocked props.

I tried many different ways to fix that, but that was the best solution.
Maybe some time I will find a better way.

If there are any problems with that fix, please let me know!
Thanks