Improved things and updated code design:
Ok. This commit is very big and concentrates on performance and code
design as well as removing stupid things.
The most things were changed in the settings. But everything should work
as it did before.

This commit is not only for us, it is also a commit for everyone,
because it makes some settings more clearly than before. Also it should
be now more userfriendly if you use PatchProtect for the first time.

So let me list the changes in detail:
- changed config names (e.g. use to enabled, ...)
- changed some names of the net-messages
- changed some strings in the panel to make settings more clearly (not
the confusing: "noantiadmin" -> just "ignore admin")
- changed some default settings values
- increased the performance of the hud a bit
- code design improvements for a better overview and with more spaces!!!
- moved some parts of the code to give the files a better structure.

Of course it could happen, that there is a new bug, but such thing
happen. Just write an issue and we will fix it ASAP