Finished the new SQL-Method
Finally I finished the new SQL-Method. There are no more ConVars used in
this version and it seems, that this buffer overflow gone too.

So here is a detailed list of the changes:
- added some code again from the removed files (mostly files like
propprotection.lua, antispam.lua), where not much changed
- changed the settings method nearly completely (no worries patcher, the
hardest parts are still the same than before and not recoded) ;)
- made a better overview in most of the files
- better performance!
- complete new antispamed-tools function
- fixed some bugs
and much more I can't remember.

So my next plans would be, to test this completely out, to see if there
is some error here and there, but normally everything should be ok now.
After that I would say, that we merge this branch to the general branch
and delte this branch.