Improved code by deleting tobool and tonumber
During the change of the sql settings, I found a problem, that every
setting-type is a string, if it is loaded from the sql table. Because of
that we also needed to convert every string to a number or a boolean.
e.g.: tobool( "1" ) = true or tonumber( "0" ) = 0

that needs much place and makes the code a bit confusing.
Now we only check the settings with REAL numbers, so we don't need the
tonumber() and tobool() in most cases anymore.
The only thing which i had to add is, that it converts the strings to
numbers, before it gets to the official sv_PProtect.Settings tables...