Added FPP Mode and PlayerPickup
SuperAdmins are now able to change the HUD-Mode. If you disable the
FPP-Mode, you will have the same HUD than before. But if you enable
FPP-Mode, you will see the owner of a prop in the middle of the screen,
right under the crosshair.

Also added the PlayerPickup. It allowes Admins to pickup players.