Finished improved World-Protection
PatchProtect blocks now all things, which should be blocked. You can
enable/disable if users, so no admins, are able to press buttons or
generally handle with them.

Also fixed the HUD for World Props. If you are looking on a world prop
you see one of 2 possibilities.
- Red means, you can't touch/toolgun it.
- Green means, that you are able to touch/toolgun/use it.
- Blue means, that you are not able to touch/toolgun it, but you are
able to use the world-buttons and doors.