Huge improvements on SQL-Settings, Stopped #54
If we added a new setting, we had to reset the sql-settings table. That
was bad.
The last version of the sql-settings was pretty old and we gained some
more experience in lua. So now it only needs to add the missing setting
to the existing sql-table.

In concolusion: Your existing sql-table doesn't reset anymore if a new
setting was added in the last update!